zip docs skills using zipSkillToBuffer
