+++ scripts/lib/zip.js
@@ -1,0 +1,21 @@
+const archiver = require('archiver');
+
+/**
+ * Create a ZIP archive for a skill directory.
+ * Returns the ZIP as a Buffer.
+ */
+async function zipSkillToBuffer(skillDir) {
+    return new Promise((resolve, reject) => {
+        const chunks = [];
+        const archive = archiver('zip', { zlib: { level: 9 } });
+
+        archive.on('data', chunk => chunks.push(chunk));
+        archive.on('end', () => resolve(Buffer.concat(chunks)));
+        archive.on('error', reject);
+
+        archive.directory(skillDir, false);
+        archive.finalize();
+    });
+}
+
+module.exports = { zipSkillToBuffer };
